faq 관련 커밋 #67

+++ src/main/java/com/tsp/new_tsp_admin/api/faq/service/AdminFaqJpaService.java
@@ -93,5 +93,5 @@ public interface AdminFaqJpaService {
      * 5. 작성일       : 2022. 08. 22.
      * </pre>
      */
-    Integer deleteFaq(Integer idx) throws Exception;
+    Long deleteFaq(Long idx) throws Exception;
 }
